axiom-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==12.0
pydantic==2.9.0
httpx==0.27.0
python-dotenv==1.0.1
numpy==1.26.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.35
sentence-transformers==3.0.0
scikit-learn==1.5.0
